refactor: update import paths for consistency and clarity in GitHub issue and pull request handling
